requirements.txt
fastapi[all]
python-dotenv
groq
numpy
sentence-transformers